+++ src/components/pages/AppDetail/components/AppDetailHeader/index.tsx
@@ -18,7 +18,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { useDispatch } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
 import {
   Typography,
   OrderStatusButton,
@@ -36,6 +36,7 @@ import { useFetchDocumentByIdMutation } from 'features/apps/apiSlice'
 import CommonService from 'services/CommonService'
 import type { UseCaseType } from 'features/appManagement/types'
 import { userHasPortalRole } from 'services/AccessService'
+import type { RootState } from 'features/store'
 
 export interface AppDetailHeaderProps {
   item: AppDetails
@@ -51,19 +52,36 @@ export interface ButtonColorType {
 export default function AppDetailHeader({ item }: AppDetailHeaderProps) {
   const { t } = useTranslation()
   const dispatch = useDispatch()
+  const isDialogConfirmed = useSelector(
+    (state: RootState) => state?.dialog?.isConfirmed
+  )
+
   const { appId } = useParams()
   const [image, setImage] = useState('')
   const [fetchDocumentById] = useFetchDocumentByIdMutation()
+  const [buttonLabel, setButtonLabel] = useState(
+    t('content.appdetail.subscribe')
+  )
 
   const getStatusLabel = (subscribeStatus: string) => {
     if (subscribeStatus === SubscriptionStatus.PENDING) {
-      return t('content.appdetail.requested')
+      setButtonLabel(t('content.appdetail.requested'))
     } else if (subscribeStatus === SubscriptionStatus.ACTIVE) {
-      return t('content.appdetail.subscribed')
+      setButtonLabel(t('content.appdetail.subscribed'))
     } else {
-      return t('content.appdetail.subscribe')
+      setButtonLabel(t('content.appdetail.subscribe'))
     }
   }
+
+  useEffect(() => {
+    if (isDialogConfirmed) {
+      setButtonLabel(t('content.appdetail.requested'))
+    }
+  }, [isDialogConfirmed])
+
+  useEffect(() => {
+    getStatusLabel(item.isSubscribed ?? SubscriptionStatus.INACTIVE)
+  }, [])
 
   const getBtnColor = (subscribeStatus: string) => {
     let btnColor: ButtonColorType
@@ -86,10 +104,7 @@ export default function AppDetailHeader({ item }: AppDetailHeaderProps) {
         break
       default:
         btnColor = {
-          color: userHasPortalRole([
-            ROLES.SUBSCRIBE_APPS,
-            ROLES.SUBSCRIBE_SERVICE,
-          ])
+          color: userHasPortalRole(ROLES.SUBSCRIBE_APP_MARKETPLACE)
             ? 'primary'
             : 'secondary',
           background1: paletteDefinitions.buttons.darkGrey ?? '',
@@ -126,18 +141,23 @@ export default function AppDetailHeader({ item }: AppDetailHeaderProps) {
 
     return (
       <OrderStatusButton
-        label={getStatusLabel(subscribeStatus)}
+        label={buttonLabel}
         color={btnColor.color}
         buttonData={OrderStatusButtonItems}
         selectable={
           subscribeStatus === SubscriptionStatus.INACTIVE &&
-          userHasPortalRole([ROLES.SUBSCRIBE_APPS, ROLES.SUBSCRIBE_SERVICE])
+          userHasPortalRole(ROLES.SUBSCRIBE_APP_MARKETPLACE)
         }
-        onButtonClick={() =>
-          subscribeStatus === SubscriptionStatus.INACTIVE &&
-          userHasPortalRole([ROLES.SUBSCRIBE_APPS, ROLES.SUBSCRIBE_SERVICE]) &&
-          dispatch(show(OVERLAYS.APPMARKETPLACE_REQUEST, appId))
-        }
+        onButtonClick={() => {
+          if (buttonLabel === t('content.appdetail.requested')) {
+            return
+          }
+          return (
+            subscribeStatus === SubscriptionStatus.INACTIVE &&
+            userHasPortalRole(ROLES.SUBSCRIBE_APP_MARKETPLACE) &&
+            dispatch(show(OVERLAYS.APPMARKETPLACE_REQUEST, appId))
+          )
+        }}
       />
     )
   }