+++ src/components/pages/CompanyData/components/EditForm.tsx
@@ -35,6 +35,7 @@ import {
   type CompanyDataType,
   useUpdateCompanySiteAndAddressMutation,
   type CompanyDataFieldsType,
+  useUpdateCompanyStatusToReadyMutation,
 } from 'features/companyData/companyDataApiSlice'
 import { useSelector } from 'react-redux'
 import {
@@ -67,6 +68,7 @@ export default function EditForm({
   const [loading, setLoading] = useState<boolean>(false)
   const [isValid, setIsValid] = useState<boolean>(false)
   const [updateData] = useUpdateCompanySiteAndAddressMutation()
+  const [updateReadyState] = useUpdateCompanyStatusToReadyMutation()
   const companyData = useSelector(companyDataSelector)
   const [success, setSuccess] = useState<boolean>(false)
   const [error, setError] = useState<boolean>(false)
@@ -79,6 +81,7 @@ export default function EditForm({
   })
   const [input, setInput] = useState<CompanyDataType>(companyDataInitialData)
   const inputParams = cloneDeep(newForm ? companyDataInitialData : companyData)
+
   if (companyInfo) {
     inputParams.externalId = `${companyInfo?.bpn}_${new Date().toISOString()}`
     inputParams.legalEntity.legalEntityBpn = companyInfo?.bpn
@@ -120,9 +123,11 @@ export default function EditForm({
     }
   }
 
-  const handleCreation = async () => {
+  const updateStateToReady = async (response: CompanyDataType[]) => {
     try {
-      await updateData([input])
+      await updateReadyState({
+        externalIds: [response[0].externalId],
+      })
         .unwrap()
         .then(() => {
           setSuccess(true)
@@ -131,6 +136,19 @@ export default function EditForm({
       setError(true)
     }
     setLoading(false)
+  }
+
+  const handleCreation = async () => {
+    try {
+      await updateData([input])
+        .unwrap()
+        .then((response) => {
+          updateStateToReady(response)
+        })
+    } catch (e) {
+      setError(true)
+      setLoading(false)
+    }
   }
 
   const getTitle = () => {