+++ src/components/pages/TechnicalUserManagement/TechnicalUserTable.tsx
@@ -21,11 +21,13 @@
 import {
   IconButton,
   PageLoadingTable,
+  StatusTag,
 } from '@catena-x/portal-shared-components'
 import { useTranslation } from 'react-i18next'
 import ArrowForwardIcon from '@mui/icons-material/ArrowForward'
 import {
   type ServiceAccountListEntry,
+  ServiceAccountStatus,
   ServiceAccountStatusFilter,
   useFetchServiceAccountListQuery,
 } from 'features/admin/serviceApiSlice'
@@ -40,6 +42,13 @@ interface FetchHookArgsType {
   statusFilter: string
   expr: string
 }
+type StatusTagColor =
+  | 'pending'
+  | 'confirmed'
+  | 'declined'
+  | 'label'
+  | 'deleted'
+  | undefined
 
 export const TechnicalUserTable = () => {
   const { t } = useTranslation()
@@ -73,11 +82,6 @@ export const TechnicalUserTable = () => {
       onButtonClick: setView,
     },
     {
-      buttonText: t('content.usermanagement.technicalUser.tabs.inactive'),
-      buttonValue: ServiceAccountStatusFilter.INACTIVE,
-      onButtonClick: setView,
-    },
-    {
       buttonText: t('content.usermanagement.technicalUser.tabs.managed'),
       buttonValue: ServiceAccountStatusFilter.MANAGED,
       onButtonClick: setView,
@@ -88,6 +92,14 @@ export const TechnicalUserTable = () => {
       onButtonClick: setView,
     },
   ]
+
+  const statusColorMap: Record<ServiceAccountStatus, StatusTagColor> = {
+    [ServiceAccountStatus.ACTIVE]: 'confirmed',
+    [ServiceAccountStatus.INACTIVE]: 'declined',
+    [ServiceAccountStatus.DELETED]: 'deleted',
+    [ServiceAccountStatus.PENDING]: 'pending',
+    [ServiceAccountStatus.PENDING_DELETION]: 'pending',
+  }
 
   return (
     <div style={{ paddingTop: '30px' }}>
@@ -126,7 +138,7 @@ export const TechnicalUserTable = () => {
           {
             field: 'serviceAccountType',
             headerName: t('global.field.type'),
-            flex: 1,
+            flex: 1.2,
           },
           {
             field: 'offer',
@@ -138,9 +150,22 @@ export const TechnicalUserTable = () => {
           {
             field: 'isOwner',
             headerName: t('global.field.owner'),
-            flex: 1,
+            flex: 0.8,
             valueGetter: ({ row }: { row: ServiceAccountListEntry }) =>
               row.isOwner ? 'Yes' : 'No',
+          },
+          {
+            field: 'status',
+            headerName: t('global.field.status'),
+            flex: 1.2,
+            renderCell: ({ row }: { row: ServiceAccountListEntry }) => (
+              <StatusTag
+                color={statusColorMap[row.status]}
+                label={t(
+                  `content.usermanagement.technicalUser.status.${row.status}`
+                )}
+              />
+            ),
           },
           {
             field: 'details',