+++ src/features/companyData/companyDataApiSlice.tsx
@@ -165,7 +165,7 @@ export interface CompanyDataType {
       deliveryServiceType: string | undefined | null
       deliveryServiceQualifier: string | undefined | null
       deliveryServiceNumber: string | undefined | null
-    }
+    } | null
     confidenceCriteria: {
       sharedByOwner: true
       checkedByExternalDataSource: true
@@ -211,15 +211,32 @@ export enum SharingStateStatusType {
   Initial = 'Initial',
 }
 
+export interface ReadyStateRequestBody {
+  externalIds: string[]
+}
+
+interface SharingStateRequest {
+  page: number
+}
+
+enum TAGS {
+  SHARING = 'sharing',
+}
+
 export const apiSlice = createApi({
   reducerPath: 'rtk/companyData',
   baseQuery: fetchBaseQuery(apiBpdmGateQuery()),
+  tagTypes: [TAGS.SHARING],
   endpoints: (builder) => ({
-    fetchSharingState: builder.query<SharingStateResponse, void>({
-      query: () => ({
-        url: '/sharing-state?page=0&size=100',
-      }),
-    }),
+    fetchSharingState: builder.query<SharingStateResponse, SharingStateRequest>(
+      {
+        query: (obj) => ({
+          url: `/sharing-state?page=${obj.page}&size=100`,
+        }),
+        keepUnusedDataFor: 5,
+        providesTags: [TAGS.SHARING],
+      }
+    ),
     fetchInputCompanyBusinessPartners: builder.mutation<
       CompanyDataResponse,
       string[] | void
@@ -241,7 +258,7 @@ export const apiSlice = createApi({
       }),
     }),
     updateCompanySiteAndAddress: builder.mutation<
-      CompanyDataResponse,
+      CompanyDataType[],
       CompanyDataType[]
     >({
       query: (data: CompanyDataType[]) => ({
@@ -249,6 +266,32 @@ export const apiSlice = createApi({
         method: 'PUT',
         body: data,
       }),
+    }),
+    updateCompanyStatusToReady: builder.mutation<
+      CompanyDataType[],
+      ReadyStateRequestBody
+    >({
+      query: (data) => ({
+        url: '/sharing-state/ready',
+        method: 'POST',
+        body: data,
+      }),
+    }),
+    uploadCSV: builder.mutation<CompanyDataType[], File>({
+      query: (body) => {
+        const formData = new FormData()
+        formData.append('file', body)
+        formData.append('type', body.type)
+        return {
+          url: '/input/partner-upload-process',
+          method: 'POST',
+          body: formData,
+        }
+      },
+      invalidatesTags: [TAGS.SHARING],
+    }),
+    downloadCsv: builder.query<void, void>({
+      query: () => '/input/partner-upload-template',
     }),
   }),
 })
@@ -258,4 +301,7 @@ export const {
   useFetchInputCompanyBusinessPartnersMutation,
   useFetchOutputCompanyBusinessPartnersMutation,
   useUpdateCompanySiteAndAddressMutation,
+  useUpdateCompanyStatusToReadyMutation,
+  useUploadCSVMutation,
+  useDownloadCsvQuery,
 } = apiSlice