+++ src/types/Config.tsx
@@ -94,6 +94,7 @@ import {
   userHasSemanticHubRole,
   userHasSsiCredentialRole,
 } from 'services/AccessService'
+import OnboardingServiceProvider from 'components/pages/OnboardingServiceProvider/OnboardingServiceProvider'
 
 /**
  * ALL_PAGES
@@ -256,7 +257,7 @@ export const ALL_PAGES: IPage[] = [
     element: <ServiceSubscription />,
   },
   {
-    name: PAGES.ADMINBOARD,
+    name: PAGES.APP_ADMIN_BOARD,
     allowTo: () =>
       userHasPortalRole([ROLES.APPROVE_APP_RELEASE, ROLES.DECLINE_APP_RELEASE]),
     element: <AdminBoard />,
@@ -271,14 +272,14 @@ export const ALL_PAGES: IPage[] = [
     element: <ServiceAdminBoard />,
   },
   {
-    name: PAGES.ADMINBOARD_DETAIL,
+    name: PAGES.APP_ADMIN_BOARD_DETAIL,
     allowTo: () =>
       userHasPortalRole([ROLES.APPROVE_APP_RELEASE, ROLES.DECLINE_APP_RELEASE]),
     isRoute: true,
     element: (
       <Route
-        key={PAGES.ADMINBOARD_DETAIL}
-        path={PAGES.ADMINBOARD_DETAIL}
+        key={PAGES.APP_ADMIN_BOARD_DETAIL}
+        path={PAGES.APP_ADMIN_BOARD_DETAIL}
         element={<AdminBoardDetail />}
       >
         <Route index element={null} />
@@ -596,6 +597,10 @@ export const ALL_PAGES: IPage[] = [
     allowTo: () => userHasPortalRole(ROLES.MY_ORGANIZATION_VIEW),
     element: <CompanyData />,
   },
+  {
+    name: PAGES.MANAGEMENT_ONBOARDING_SERVICE_PROVIDER,
+    element: <OnboardingServiceProvider />,
+  },
 ]
 
 export const ALL_OVERLAYS: RestrictedItem[] = [
@@ -738,6 +743,9 @@ export const ALL_OVERLAYS: RestrictedItem[] = [
   {
     name: OVERLAYS.COMPANY_CERTIFICATE_CONFIRM_DELETE,
   },
+  {
+    name: OVERLAYS.CSV_UPLOAD_OVERLAY,
+  },
 ]
 
 export const ALL_ACTIONS: IAction[] = [
@@ -789,7 +797,7 @@ export const mainMenuFullTree = [
       { name: PAGES.APP_OVERVIEW, hint: HINTS.NEW },
       { name: PAGES.APP_RELEASE_PROCESS },
       { name: PAGES.APP_SUBSCRIPTION, hint: HINTS.NEW },
-      { name: PAGES.ADMINBOARD, hint: HINTS.NEW },
+      { name: PAGES.APP_ADMIN_BOARD, hint: HINTS.NEW },
     ],
   },
   {
@@ -829,6 +837,7 @@ export const userMenuFull = [
   PAGES.COMPANY_CERTIFICATE,
   PAGES.COMPANY_WALLET,
   PAGES.COMPANY_DATA,
+  PAGES.MANAGEMENT_ONBOARDING_SERVICE_PROVIDER,
   PAGES.LOGOUT,
 ]